Collapse some client clogic. Verification should be done on server

+++ client/birdComm/birdComm.go
@@ -15,44 +15,37 @@ import (
 
 func main() {
 
-	// TO-DO
-	// Option to add/delete static routes
-
 	// Get new peer data
 	server := flag.String("server", "localhost", "Config server")
 	name := flag.String("name", "", "new peer name")
 	desc := flag.String("description,", "", "description of peer")
 	action := flag.String("action", "", "action to perform")
-	// TO-DO
-	// Should use an IP library to ensure entered IP is valid
 	address := flag.String("address", "", "address of peer")
 	as := flag.Uint("as", 0, "as number of peer")
 	prefix := flag.String("prefix", "", "A prefix to get to")
 	mask := flag.Uint("mask", 0, "a subnet mask")
 	nexthop := flag.String("nexthop", "", "A nexthop")
+
 	flag.Parse()
 
 	if *action == "" {
 		log.Fatalf("Need an action to perform")
 	}
 
-	// We need a certain minimum of data to add a neighbour
-	/*if *name == "" || *address == "" || *as == 0 {
-		log.Fatalln("At minimum, address, as, and name is required")
-	}*/
-
 	// TO-DO
 	// Need to be able to specify address family.... Should also ensure
 	// it's valid
 
+	// Set up connection to server
 	serverConn := fmt.Sprintf("%v:1179", *server)
 	conn, err := grpc.Dial(serverConn, grpc.WithInsecure())
 	if err != nil {
 		log.Fatalf("Failed to start gRPC connect: %v", err)
 	}
 	defer conn.Close()
-
 	client := pb.NewBirdCommClient(conn)
+
+	// Fill peer and route messages with parsed data
 	peer := &pb.Peer{
 		Name:        *name,
 		Description: *desc,
@@ -69,26 +62,26 @@ func main() {
 	}
 
 	switch *action {
-	case "DeleteNeighbour":
-		res, err := delete(peer, client)
+	case "AddPeer":
+		res, err := peerAction(peer, client, true)
 		if err != nil {
 			log.Fatalf("error received: %v\n", err)
 		}
 		fmt.Printf("%v\n", res)
-	case "AddNeighbour":
-		res, err := add(peer, client)
+	case "DeletePeer":
+		res, err := peerAction(peer, client, false)
 		if err != nil {
 			log.Fatalf("error received: %v\n", err)
 		}
 		fmt.Printf("%v\n", res)
 	case "AddRoute":
-		res, err := addS(route, client)
+		res, err := routeAction(route, client, true)
 		if err != nil {
 			log.Fatalf("error received: %v\n", err)
 		}
 		fmt.Printf("%v\n", res)
 	case "DeleteRoute":
-		res, err := deleteS(route, client)
+		res, err := routeAction(route, client, false)
 		if err != nil {
 			log.Fatalf("error received: %v\n", err)
 		}
@@ -98,7 +91,17 @@ func main() {
 	}
 }
 
-func delete(p *pb.Peer, client birdComm.BirdCommClient) (*pb.Result, error) {
+func peerAction(p *pb.Peer, client birdComm.BirdCommClient, add bool) (*pb.Result, error) {
+	// Add peer if it's new
+	if add {
+		resp, err := client.AddNeighbour(context.Background(), p)
+		if err != nil {
+			log.Fatalf("Received an error from gRPC server: %v", err)
+		}
+		return resp, err
+	}
+
+	// delete peer otherwise
 	resp, err := client.DeleteNeighbour(context.Background(), p)
 	if err != nil {
 		log.Fatalf("Received an error from gRPC server: %v", err)
@@ -106,23 +109,17 @@ func delete(p *pb.Peer, client birdComm.BirdCommClient) (*pb.Result, error) {
 	return resp, err
 }
 
-func add(p *pb.Peer, client birdComm.BirdCommClient) (*pb.Result, error) {
-	resp, err := client.AddNeighbour(context.Background(), p)
-	if err != nil {
-		log.Fatalf("Received an error from gRPC server: %v", err)
+func routeAction(r *pb.Route, client birdComm.BirdCommClient, add bool) (*pb.Result, error) {
+	// Add route if it's new
+	if add {
+		resp, err := client.AddStatic(context.Background(), r)
+		if err != nil {
+			log.Fatalf("Received an error from gRPC server: %v", err)
+		}
+		return resp, err
 	}
-	return resp, err
-}
 
-func addS(r *pb.Route, client birdComm.BirdCommClient) (*pb.Result, error) {
-	resp, err := client.AddStatic(context.Background(), r)
-	if err != nil {
-		log.Fatalf("Received an error from gRPC server: %v", err)
-	}
-	return resp, err
-}
-
-func deleteS(r *pb.Route, client birdComm.BirdCommClient) (*pb.Result, error) {
+	// delete route otherwise
 	resp, err := client.DeleteStatic(context.Background(), r)
 	if err != nil {
 		log.Fatalf("Received an error from gRPC server: %v", err)
